[INCOME-2935] Optimizar borrado PIPE (#1091)
* feat(INCOME-2935): implement pagination for commission migration and add person count query

* test(INCOME-2935): add SQL_COUNT_PERSONAS constant and update tests for migration services

* ci(INCOME-2935): add pipeborrado slot

* refactor(INCOME-2935): update parameter names in SQL query

* [INCOME-2935] Modificacion formato TIPO_PORCENTAJE_POLITICA en consulta envio de datos a PIPE

* [INCOME-2935] Modificacion formato TIPO_PORCENTAJE_POLITICA en consulta envio de datos a PIPE

* feat(INCOME-2935): config jms for slot

* [INCOME-2935] Modificacion PAGE-SIZE ed migracion comision para prueba de carga

* [INCOME-2935] Modificacion PAGE-SIZE en migracion comision a valor estandar. Añadir Trabajo y Tarea a logs. Cambio de visivilidad a private del metodo conteo de personas

* [INCOME-2935] Eliminar slot en deployments

* [INCOME-2935] Condicionar el envío a PIPE en las simulaciones

* [INCOME-2935] Tests

---------

+++ code/APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/app/run/tarea/service/RunTareaServiceImpl.java
@@ -113,13 +113,19 @@ public class RunTareaServiceImpl implements RunTareaService {
         this.runTareaLimpiarConsolidarByAmbitoService.run(runTarea);
       }
 
-      final List<TareaMigrarComisionDto> deleteMigracion = this.tareaMigrarService.deleteCalculoComisionByTareaActual(runTarea,
-          runTarea.getTarea().getAmbito().get(0));
+      if (runTarea.getTrabajo().getIdSimulacion() == null) {
+        final List<TareaMigrarComisionDto> deleteMigracion = this.tareaMigrarService.deleteCalculoComisionByTareaActual(runTarea,
+            runTarea.getTarea().getAmbito().get(0));
+        this.runTareaConsolidarService.run(runTarea);
+        this.tareaService.updateEstadoFinal(runTarea.getTarea());
+        this.tareaService.updateFechaFin(runTarea.getTarea());
+        this.runTareaMigrarService.run(runTarea, deleteMigracion);
+      } else {
+        this.runTareaConsolidarService.run(runTarea);
+        this.tareaService.updateEstadoFinal(runTarea.getTarea());
+        this.tareaService.updateFechaFin(runTarea.getTarea());
+      }
 
-      this.runTareaConsolidarService.run(runTarea);
-      this.tareaService.updateEstadoFinal(runTarea.getTarea());
-      this.tareaService.updateFechaFin(runTarea.getTarea());
-      this.runTareaMigrarService.run(runTarea, deleteMigracion);
     } catch (final ValidationNoReintentoException | ValidationReintentoException e) {
       if (e instanceof ValidationNoReintentoException) {
         this.tareaCalculoPersonaService.updateWithEstado(runTarea, EstadoTareaCalculoPersonaEnum.PENDIENTE.getDto(),
